dom/base/test/chrome/test_bug1098074_throw_from_ReceiveMessage.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1098074
-->
<window title="Mozilla Bug 1098074"
    xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
    onload="start();">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  /** Test for Bug 1098074 **/
  SimpleTest.waitForExplicitFinish();
  SimpleTest.expectUncaughtException();

  // Tell the test to expect exactly one console error with the given parameters,
  // with SimpleTest.finish as a continuation function.
  SimpleTest.monitorConsole(SimpleTest.finish, [{errorMessage: new RegExp('acopia')}]);

  var globalMM = Cc["@mozilla.org/globalmessagemanager;1"].getService();
  globalMM.addMessageListener("flimfniffle", function onMessage(msg) {
    globalMM.removeMessageListener("flimfniffle", onMessage);
    is(msg.data, "teufeltor", "correct message");

    // Cleanup the monitor after we throw.
    SimpleTest.executeSoon(SimpleTest.endMonitorConsole);

    // eslint-disable-next-line no-throw-literal
    throw "acopia";
  });

  function start() {
    globalMM.loadFrameScript("data:,sendAsyncMessage('flimfniffle', 'teufeltor')", true);
  }

  ]]>
  </script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1098074"
     target="_blank">Mozilla Bug 1098074</a>
  </body>
</window>
